Prevent segfault when xcb_icccm_get_wm_icon_name_reply() fails.
Chormium was the application that I saw this with; for whatever reason, it causes xcb_icccm_get_wm_icon_name_reply() to fail.  If the xcb_icccm_get_wm_icon_name_reply() call returns an error, prop.name is not a valid pointer and should not be dereferenced.  At best it points to the last valid request, at worse it causes a seg-fault.

This is not an ideal solution but I was having difficulty finding any reasonable documentation on how/why xcb_icccm_get_wm_icon_name_reply() might fail and what to do in that case.  This at least avoids the seg fault and puts something in the blank spot we'd otherwise have.