Start of slow cleanup and move to ewmh functions
This change move some of the ewmh hints to the functions available in
the xcb ewmh extension provided and already included in 2bwm.

This is the start of a slow cleanup of the code in the hope that new
highly wanted features will be more easily implemented and won't clutter
the ideology and the codebase.