Signal handling switch to sigaction(2) instead of signal(2)
Switch the signal call to sigaction, as signal(2) is deprecated.
Modified Makefile because `-std=c99` didn't support sigaction, we'll see
how this goes with everyone.
Added support for SIGHUP signal handling so that 2bwm restart upon
receiving it.
Mentioning how signals are handled in the manpage (more info the better)