Fix crash caused by hardcoded TWOBWM_PATH
This fixes a potential crash due to hardcoding TWOBWM_PATH in the
Makefile. If a user does not install the 2bwm binary to the hardcoded
path, as can happen with certain package managers, 2bwm will crash when
calling twobwm_restart. Instead, a global variable is set to argv[0]
which contains the program name and the 'p' flavor of exec syscalls will
find 2bwm in the user's $PATH. This commit also fixes a warning
triggered by -Wnonnull where execvp was receiving NULL instead of its
expected vector argument. The call was changed to execlp and supplied
the expected arguments.